OSpeaker photo upload fixed
