+++ app/helpers/wizard/sponsors.py
@@ -2,8 +2,7 @@ from app.helpers.data import save_to_db
 from app.helpers.data_getter import DataGetter
 from app.helpers.helpers import represents_int
 from app.helpers.storage import UPLOAD_PATHS
-from app.helpers.wizard.helpers import save_event_image
-from app.models import db
+from app.helpers.assets.images import save_event_image
 from app.models.sponsor import Sponsor
 
 
@@ -16,7 +15,7 @@ def get_sponsors_json(event_id_or_sponsors):
     data = []
     for sponsor in sponsors:
         data.append(sponsor.serialize)
- 
+
     return sorted(data, key=lambda x: x['id'])
