+++ app/templates/gentelella/admin/event/sessions/components/_speaker_croppie.html
@@ -34,7 +34,13 @@
         e.unwrap();
     }
     $(document).ready(function () {
-        $uploadCropper = $('#upload-cropper').croppie({
+
+        var $photoUploadInput = $("#speaker-photo-upload");
+        var $uploaderLabel = $("#speaker-photo-upload-label");
+        var $cropperModel = $('#cropper-modal');
+        var $photoUrlInput = $("#photo");
+
+        var $uploadCropper = $('#upload-cropper').croppie({
             viewport: {
                 width: 190,
                 height: 190,
@@ -46,7 +52,7 @@
             }
         });
 
-        $("#event-image-upload").unbind("change").on('change', function () {
+        $photoUploadInput.unbind("change").on('change', function () {
             var input = this;
             if (input.files && input.files[0]) {
                 if (input.files[0].size > 10485760) {
@@ -66,124 +72,60 @@
             else {
                 logError("No FileReader support");
             }
-
         });
 
-        $("#logo-upload").unbind("change").on('change', function () {
-            var input = this;
-            if (input.files && input.files[0]) {
-                if (input.files[0].size > 1048576) {
-                    createSnackbar("Image must be less than 1 MB in size");
-                    resetFormElement(input);
+        var loadingImage = '<img src="' + "{{ url_for('static', filename='img/loading.gif') }}" + '" style="width: 50px;"> <h3>Uploading</h3>';
+
+        $("#save-crop").click(function () {
+            $uploadCropper.croppie('result', {
+                type: 'canvas',
+                size: 'original'
+            }).then(function (resp) {
+                $cropperModel.modal('hide');
+                $uploaderLabel.html(loadingImage);
+                $.ajax({
+                    type: 'POST',
+                    url: '/events/create/files/image/',
+                    data: {
+                        image: resp
+                    },
+                    dataType: 'json'
+                }).done(function (data) {
+                    $("#image-view-group").show().find("img").attr("src", resp);
+                    $("#image-upload-group").hide();
+                    $photoUrlInput.val(data.image_url);
+                }).fail(function (data) {
+                    createSnackbar("An error occurred. Please try again later.");
+                }).always(function () {
+                    $uploaderLabel.html('<i class="fa fa-3x fa-camera-retro" aria-hidden="true"></i><br><h3>Select Avatar</h3>');
+                });
+            });
+        });
+
+        $("#remove-image-btn").click(function () {
+            $.ajax({
+                type: 'DELETE',
+                url: "{{ url_for('event_speakers.avatar_delete', event_id=event.id, speaker_id=speaker.id) }}",
+                dataType: 'json'
+            }).done(function (data) {
+                if (data.hasOwnProperty('status') && data.status == 'ok') {
+                    $("#image-view-group").hide();
+                    $("#image-upload-group").show();
+                    $photoUrlInput.val('');
                 } else {
-                    var reader = new FileReader();
-                    reader.onload = function (e) {
-                        resetFormElement(input);
-                        $("#logo").val(e.target.result);
-                        $("#logo-view-group").show().find("img").attr("src", e.target.result);
-                        $("#logo-upload-group").hide();
-                    };
-                    reader.readAsDataURL(input.files[0]);
+                    createSnackbar("An error occurred. Please try again later.");
                 }
-            }
-            else {
-                logError("No FileReader support");
-            }
-        });
-
-        $("#remove-logo-btn").click(function () {
-            $("#logo-view-group").hide();
-            $("#logo-upload-group").show();
-            $("#logo").val('');
-        });
-
-        function resetFormElement(e) {
-            e = $(e);
-            e.wrap('<form>').closest('form').get(0).reset();
-            e.unwrap();
-        }
-    });
-
-    var loadingImage = '<img src="' + "{{ url_for('static', filename='img/loading.gif') }}" + '" style="width: 50px;"> <h3>Uploading</h3>';
-    $("#save-crop").click(function () {
-        $uploadCropper.croppie('result', {
-            type: 'canvas',
-            size: 'original'
-        }).then(function (resp) {
-            $('#cropper-modal').modal('hide')
-            $("#event-image-upload-label").html(loadingImage);
-            /* Disable Save Event buttons */
-            $(".save-event").prop("disabled", true);
-            $(".publish-unpublish-event").prop("disabled", true);
-            $.ajax({
-                type: 'POST',
-                url: "{{ url_for('event_speakers.bgimage_upload', event_id=event.id, speaker_id=speaker.id) if speaker.photo else url_for('event_speakers.create_event_bgimage_upload', event_id=event.id) }}",
-                data: {bgimage: resp},
-                dataType: 'json'
-            }).done(function (data) {
-                console.log(data);
-                $("#logo").val(data.background_url);
-                $("#image-view-group").show().find("img").attr("src", resp);
-                $("#image-upload-group").hide();
             }).fail(function (data) {
-                alert("Something went wrong. Please try again.");
-            }).always(function () {
-                $("#event-image-upload-label").html('<i class="fa fa-3x fa-camera-retro" aria-hidden="true"></i><br><h3>Select Avatar</h3>');
-                /* Enable Save Event buttons */
-                $(".save-event").prop("disabled", false);
-                $(".publish-unpublish-event").prop("disabled", false);
-            });
-        });
-    });
-
-    {% if profile.user_detail.avatar_uploaded %}
-        $("#remove-image-btn").click(function () {
-            $.ajax({
-                type: 'DELETE',
-                url: "{{ url_for('event_speakers.bgimage_upload', event_id=event.id, speaker_id=speaker.id) }}",
-                dataType: 'json'
-            }).done(function (data) {
-                console.log(data);
-                $("#image-view-group").hide();
-                $("#image-upload-group").show();
-                $("#logo").val('');
-            }).fail(function (data) {
-                alert("Something went wrong. Please try again.");
+                createSnackbar("An error occurred. Please try again later.");
             });
         });
 
-        $("#remove-logo-btn").click(function () {
-            $.ajax({
-                type: 'DELETE',
-                url: "{{ url_for('events.logo_upload', event_id=event.id) }}",
-                dataType: 'json'
-            }).done(function (data) {
-                console.log(data);
-                $("#logo-view-group").hide();
-                $("#logo-upload-group").show();
-                $("#logo").val('');
-            }).fail(function (data) {
-                alert("Something went wrong. Please try again.");
-            });
+        $cropperModel.on('shown.bs.modal', function (e) {
+            $uploadCropper.croppie('bind');
+            resetFormElement($photoUploadInput);
         });
-    {% else %}
-        $("#remove-image-btn").click(function () {
-            $("#image-view-group").hide();
-            $("#image-upload-group").show();
-            $("#logo").val('');
-        });
-
-        $("#remove-logo-btn").click(function () {
-            $("#logo-view-group").hide();
-            $("#logo-upload-group").show();
-            $("#logo").val('');
-        });
-    {% endif %}
-
-    $('#cropper-modal').on('shown.bs.modal', function (e) {
-        $uploadCropper.croppie('bind');
-        resetFormElement($("#event-image-upload"));
 
     });
+
 
 </script>
